Eilmer: make the MappedCellCopy class safe to use in an MPI simulation.
To we have taken the minimum-effort approach and added just enough code to
make the use of this GhostCellEffect safe to use, by restricting potential
ghost-cell-to-source-cell mappings to be within a single MPI process.

This restriction will be removed once the FullFaceCopy is running properly
with MPI communication, but that is not today's job.