eilmer/examples: 3D connection test is now tolerant of small floating point differences.
The 3D connection test started failing somewhere along the line. Possibly, this was due
to changes in the details of optimisations the compiler makes. I had been testing for
errors by doing a diff on the string output. This was a very exacting test, and overly
sensitive.

In the latest commit, I've made use of Peter's fpdiff module to compare the output
based on floating point differences. It is now tolerant of small differences in
floating point values. The connection tests all pass using this approach.