nm: modified method for adding values to 0 entries in sparse matrix class.
I have (hopefully) improved the efficiency of adding values to 0
entries in the sparse matrix class by using the D language array library
algorithm, insertInPlace().

This update to the smla.d file also slightly improves the memory costs
of the ILUp level-based pre-conditioner. This is achieved by only allowing
non-zero entries in the sparse matrix to be updated with a 0.0 value. Currently
we have chosen ESSENTIALL_ZERO to be 1.0e-50.