eilmer: removed residual sensitivity w.r.t. mesh points, reworked adjoint solver
The initial implementation of the adjoint solver followed a method that required
the formation of not only the flow sensitivity matrix, but also the residual
sensitivity w.r.t mesh points matrix (in addition to the number of other smaller
sensitivity matrices). The large number of finite-difference parameters made
verification tedious, and difficult.

The updated implementation removes the residual sensitivity w.r.t. mesh points,
by computing the residual sensitivity w.r.t design variables in one step (see
the literature for the mathematics). This removes some of the finite-difference
parameters, and makes for a simplified implementation.

Note that if we wish to implement a mesh adjoint, then we will need to return
to the formation of the residual sensitivity w.r.t. mesh point matrix.