eilmer: Changed default gradient estimator and put in consistency checks.
The spatial gradient estimator defaults to "divergence" at "vertices".
However, if a 3 simulation or a simulation containing an unstructured grid
is detected, then the gradient estimator selection is changed to "least_squares"
at "faces". The user is issued a Note when this change is made. The .config
file will have the selection recorded.