Eilmer: In the block-marching mode, fix the propagation of upstream data.
When the MPI code was implemented (back in rev 1503, 2018-01-24),
the actual copying of boundary-cell data was shifted out of
the ghost-cell-effect action and into dedicated exchange functions.
Now, to get the upstream flow data into the newly-activated blocks'
ghost cells, we need to call the top-level exchange function.

After propagating the data across the block, we do a more complete
initialization of the conserved quantities before resuming time-integration.