grid_utils: weighting for the IDW now uses original surface vertex position.
Originally we were computing the weighting with the new vertex position, which
worked well for small perturbations of the boundaries. For large perturbations,
poor grid quality was noted, and in extreme cases broken meshes (inverted cells)
occured. It has been found that using the original vertex position in the
weighting calculation retains grid quality much better.