nm: removed assert from the sparse matrix-vector product function.
We initially had an assert in the sparse matrix-vector product that
assumed the given matrix is square. However, it is not a fundamental
requirement of the Compressed Row Storage (CRS) format, nor of the
sparse matrix-vector product, that the given matrix be square. I have
thus removed the assert. As I have noted in the file, this now puts the
onus on anyone using this method to ensure that the vector length is
the same length as the number of columns in the sparse matrix, since this
knowledge is not easily determined in CRS.