eilmer: removed mapped cell gradient copy from shape sensitivity perturbed flux routine.
Consider the formation of the Jacobian in parallel for a domain decomposed into multiple blocks.
When we perturb a cell along the boundary, this in effect should change the gradient in the ghost cell.
If using the shared-memory flavour of the flow solver we should technically copy across the perturbed cell
value to the neighbour blocks ghost cell, perform the gradient computation for the mapped cell, and transfer it
back to the original (local) block for the flux computation. This would be inefficient, and problematic when
considering that the neighbour block is also currently forming it's Jacobian contribution.

To overcome this we will currently just use the fall back for the numerical Jacobian formation, i.e., we will
just copy the gradient from the other side of the interface of a ghost cell (in other words we just copy the perturbed
cells gradient into the ghost cell).