eilmer: Shifted the call to diffuseWallsIntoBlock from init_simulation() to steadystate_solver.d
diffuseWallsIntoBlock was orignally called during the simulation initialisation, which was causing
some conflicts when a simulation was being initialised for use with the shape senstivity calculator.
To remedy this, I have shifted the call to diffuseWallsIntoBlock to the steady-state solver main function.
It is executed directly after the simulation has been initialised.