eilmer: added option to read and write bezier parameterising design surface to file.
Computational time is excessive when fitting an n-point bezier curve (where n is large)
to a surface defined by a large number of vertices, hence we do not want to have to re-parameterise
the surface for each execution of the shape sensitivity calculator. When performing an optimisation of
a geometry we really only need to run the parameterisation method once as a pre-processing step and
store the bezier information for each subsequent execution of the shape senstivity calculator in a set of files.
This update adds this ability to the shape sensitivity calculator.