kinetics: Fix bugs related to parsing electrons in chemistry files.
Recently I allowed for species names that start with lowercase
letters to suit some Chemkin-style input files. However, when
making that change, I forgot to include the electron as a special
case. The electron also starts with a lowercase letter "e".
It should be treated as a special case. This is fixed in this
commit.