Eilmer: Introduce shear-stress and heat-transfer limits.
These limits are based on the idea that, in a real flow with discrete molecules,
shear stress would be limited to some fraction of the local static pressure.
The intention is to mitigate extreme values that are the result of poor least-
squares estimates of flow gradients and nasty sharp-corner flows.