Eilmer: update wall function implementation for improved robustness
The old implementation of the wall function boundary condition seemed
to cause simulations to crash whenever we have a large velocity and/or
thermal gradients near non-slip walls. This is likely to have been
brought about by two things:

  1) the way the old implementation simultaneously solves for q_wall
     and tau_wall (which creates a race condition).
  2) the asin(alpha) function which crashes whenever alpha > 1.0.

To fix this, we now solve for both q_wall and tau_wall in nested loops
(to fix issue 1), and also implemented a limiter on alpha (to fix
issue 2). The high-speed cold-wall simulations appear to run smoothly
and produce boundary layers with reasonable thicknesses with this new
implementation. Now we will just have to run some test cases to
quantitatively validate this.

More to-do for the wall function BC:

  1) As per Anand's suggestion, implement the improved version of this
     wall function as reported in Gao's paper.
  2) As per Rowan's suggestion, implement an adaptive switch which
     selects either the wall function or solve-to-wall depending on y+.