eilmer: added viscous effects back into the shape sensitivity calculator.
The flow Jacobian with viscous effects passes the complex-step Frechet derivative test (*). However,
the expected level of agreement between the gradients computed via the direct complex-step method is
not achieved. The cause is unknown, however given the positive result of the flow Jacobian, I suspect the
subtle bug is in the direct method, not the adjoint method.

(*) note that the test only passes when the viscous stress limiting lines of code in the
fvinterface.viscous_flux() method is removed.