eilmer: steady-state solver now can use ILU0 as a preconditioner.
We may now select an ILU0 decomposition of the 1st order Jacobian w.r.t.
the conservative variables, in addition to the block-diagonal preconditioner
already avaiable. The ILU0 preconditioner converges the supersonic ramp
test case (from the Marques and Pereira paper) in roughly half the amount
of steps the block-diagonal preconditioner achieves.

Note that computing the ILU0 precondition matrix can be computationally
expensive. So much so that the gain in convergence maybe out-weighed by
the increase in computational time per step. An approach that is taken from
the literature has been applied, where the user can choose to only update
the precondition matrix every n steps. In the supersonic ramp test case
above, using the ILU0 preconditioner with n = 10, convergence is achieved
in about half the amount of wall-clock time the block-diagonal preconditioner
achieved.

To select a preconditioner, set:

   precondition_matrix_type = "block_diagonal" or "ilu"

To select the frequency at which the preconditioner is updated, set:

   frozen_preconditioner_count = 10 (for example)