eilmer: added in complex-step sensitivity for source terms.
This addition allows the Jacobian to be formed for the axisymmetric solver,
which requires both inviscid and viscous source terms.

This update also holds a bug fix in the boundary condition for the conservative Jacobian;
during the transformation step from primitive to conservative form, the transform
matrix was using an incorrect cell flowstate.