examples/kinetics: Updated state-specific reactor to use 3-species, 2 reactions
The example now runs with 3 pseudo-species (N(4s), N2(v=0) and N2(v=1))
and two dissociation reactions. This gives a good comparison to the
thermal equilibrium finite-rate chemistry calculation. The addition of an
extra species and reaction gives confidence that the loops are working for
multiple species and reactions in the implementation.