Eilmer: A small optimization of one of the reconstruction functions.
When running the code with profiling instrumentation, the clip_to_limits
function was near the top of the pile of measured CPU time.

The new code is not quite so concise but it is significantly faster.
The sharp-cone-20-degrees case now runs in 5 seconds, compared with 7 seconds.

We should do more of this profiling.