Gasdyn: introduce a GasFlowException.
This is the first stage of tightening the checks on the gas fow calculations
while still allowing recovery at a higher level of the code.

The immediate application is to be able to signal and trap exceptional conditions
in the Lua scripts that might be part of an automated optimization process.
We would prefer these scripts to be able to recover from stepping into invalid
regions of parameter space.