Try to reduce the production of temporary copies of Vector3 objects.
This is the first of a few commits that might make the code a little more ugly
since it will probably replace a few Vector3 arithmetic expressions with more
explicit steps.

The starting point for this commit was an attempt to improve detect_shock_points
but the changes rippled out into other code.