Eilmer, Geom: make set_gcl_interface_properties() @nogc.
This reduces the number of dup calls in Fabs' moving grid test from 155.8 million
down to 113.3 million.  To do this we had to introduce another couple of combined
assignment operators with "const(Vector3)" type for the rhs argument, because the
compiler refused to use the "in Vector3" type in a @nogc function.

Also, we changed the signature of the combined operator methods and normalize()
to return void.  We will not be able to use them in expressions.

The next step is to make @nogc more of the geometric functions that
are called frequently in the moving grid simulations.