kinetics: Added electronically_specific kinetics module (by Brad Semple)
This commit adds the shell of Brad Semple's kinetics update for
an electronically specific gas. In the process of getting the code
to build, we also needed to add the bdfLU module to nm package
and make some changes in the gas package.

This commit compiles cleanly, but the module is not functional.