Eilmer: refactor main time-stepping loop.
Have pushed some of the code within the main time-stepping loop out into a set
of functions that, hopefully, will show up in the profiling log file.
We might then have an easier time trying to identify why the integrate_in_time
function takes such a lot of CPU time to do its work.

Along the way, the refactor cleaned up a bit of the code and made it simpler,
however, I seem to have broken the MPI flavour of the code such that the
main stepping loop finishes immediately.  The problem seems to be in the MPI
reduction of the boolean localFinishFlag across MPI tasks.  All values go in
as false but come out as true.  Look for the debug blocks either side of the
MPI_Allreduce call on line 759.

This commit is for review, not for general use.