Two temperature reacting argon model files:
gas: Have adjusted the default value of the electron temperature to be the heavy
particle temperature rather than an arbitrary but user configurable number.

kinetics: Have implemented the option of using the backward_euler integration
scheme for the chemical kinetics. At the moment it seems to work fine in the
unit test, however is crashing in Eilmer4 simulations.