Eilmer: use int variables for logical reduction across MPI tasks.
I suspect that the D compiler packs several boolean variables together such that,
just giving the address of one of them to the C domain results in confusion of
the values.  The work-around is to explicitly convert to int values in the D code
and then reduce those int values across the MPI tasks.