eilmer: parallelise ILU preconditioner for the steady-state solver.
This is the first attempt at parallelising the ILU preconditioner. It seems
to hold together nicely for the Coles flat plate (turbulent) test case.

Note: we use the additive-Schwarz with zero-overlap approach, which is a convoluted way
of saying we form an ILU precondition matrix for each block without considering the
effects of neighbouring blocks.