eilmer: Make binary diffusion coefficients available in Lua.
This commit replaces the Lua-exposed function in revision
1973. Rev 1973 made the averaged diffusion coefficient available.
However, this was not that useful for Andrew Cameron's work
on a catalytic boundary condition. Instead, the individual
binary coefficients are needed. This commit makes them accessible
to the user-defined boundary conditions using:

  computeBinaryDiffCoeffs(T, p, D)

Note the caller should already provide a D table as
an array of arrays with enough spaces available, e.g.

D = {}
for isp=0,n_species-1 do
   D[isp] = {}
   for jsp=0,n_species-1 do
      D[isp][jsp] = 0.0
   end
end