eilmer: added freeze_limiter_on_step to globalconfig.
Despite the flow solution being converged with regards to surface integral
quantities (lift, drag, heat transfer), benign rining in the unstructured
limiters prevents a sufficient drop in the residuals for use with adjoint
and direct gradient methods. A reported method for fixing this known
issue is to freeze the limiters after a set number of steps. This revision
adds the user input variable for setting the step at which the limiter is
frozen -- however the mechanics of the freezing will be added in a later
revision.