Eilmer: @nogc experiment of putting Exception throwing into debug blocks.
Just for a single function at the moment...
If we put any actions that require allocations into debug blocks,
the DMD compiler will allow the function to be marked @nogc.
Now we can  make many more functions @nogc to assist with
our campaign to eliminate unnecessary temporary copies.
There is a compromise, however, in that non-debug versions of the code
will just use asserts which provide less information and are not catchable.