eilmer: added parameters that allow the user to tweak the CFL ramp for the steady-state solver.
We use the Switched Evolution Relaxation (SER) to increase the timestep (i.e. ramp the CFL) for
the steady-state solver. The SER method increases the timestep based on the inverse of the residual
reduction raised to a power. Initially we had hard-coded powers of 0.75 and 1.0 for the initial and
inexact-Newton phases of the Newton-Krylov method. These values appear to be too agressive for some
turbulent simulations (as also noted in the literature). For this reason, I have added these parameters
to the user-configurable sssOptions. The defaults remain as 0.75 and 1.0.