Eilmer: first cut at reading flow data from a file with variable order of items
...on each line of the cell data.

At the moment, the variable-order scanning function only works for a single
thread in e4shared and fails with a nasty Segmentation fault when run with
more than one thread.  Did not yet check MPI with a single thread per MPI task.

This commit is for review by Kyle and Rowan.

Maybe the block-local gas model needs to be passed through to the massfName
function...  That's tomorrow's job.