Eilmer: get the code to determine size of time step right from the beginning.
The user may specify config.dt_init to override the time-step size if the code
seems to be determining a size that is too big to maintain numerical stability.
This commit also changes the default value of dt_init to 1.0e-3 so that the
automatic calculation of dt_init is likely to be in effect.

A few of the examples have been updated to allow the automatic calculation to
do its thing and to make the test scripts consistent with the new numbers
of steps taken in the simulations.

This commit can be blamed on Nils for asking "But why would you do it that way?"
when discussing the previous logic around the determination of time-step size.